add uml_switch header
